+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Functional research topic template
 - Added admin dashboard
   - Ability to create team members for static page
+  - Ability to create partners for static page